refactor(security): share WHIR hiding accounting
